+++ src/include/common/jsonapi.h
@@ -89,7 +89,7 @@ typedef struct JsonIncrementalState JsonIncrementalState;
 typedef struct JsonLexContext
 {
 	char	   *input;
-	int			input_length;
+	size_t		input_length;
 	int			input_encoding;
 	char	   *token_start;
 	char	   *token_terminator;
@@ -158,7 +158,7 @@ extern JsonParseErrorType pg_parse_json(JsonLexContext *lex,
 extern JsonParseErrorType pg_parse_json_incremental(JsonLexContext *lex,
 													JsonSemAction *sem,
 													char *json,
-													int len,
+													size_t len,
 													bool is_last);
 
 /* the null action object used for pure validation */
@@ -193,7 +193,7 @@ extern JsonParseErrorType json_count_array_elements(JsonLexContext *lex,
  */
 extern JsonLexContext *makeJsonLexContextCstringLen(JsonLexContext *lex,
 													char *json,
-													int len,
+													size_t len,
 													int encoding,
 													bool need_escapes);
 
@@ -219,6 +219,6 @@ extern char *json_errdetail(JsonParseErrorType error, JsonLexContext *lex);
  *
  * str argument does not need to be nul-terminated.
  */
-extern bool IsValidJsonNumber(const char *str, int len);
+extern bool IsValidJsonNumber(const char *str, size_t len);
 
 #endif							/* JSONAPI_H */
